third_party/blink/web_tests/fast/dom/MutationObserver/parser-mutations-expected.txt

PASS mutations.length is 22
PASS mutations[0].type is "childList"
PASS mutations[1].type is "childList"
PASS mutations[2].type is "childList"
PASS mutations[3].type is "childList"
PASS mutations[4].type is "childList"
PASS mutations[5].type is "childList"
PASS mutations[6].type is "childList"
PASS mutations[7].type is "childList"
PASS mutations[8].type is "childList"
PASS mutations[9].type is "childList"
PASS mutations[10].type is "childList"
PASS mutations[11].type is "childList"
PASS mutations[12].type is "childList"
PASS mutations[13].type is "childList"
PASS mutations[14].type is "childList"
PASS mutations[15].type is "childList"
PASS mutations[16].type is "childList"
PASS mutations[17].type is "childList"
PASS mutations[18].type is "childList"
PASS mutations[19].type is "childList"
PASS mutations[20].type is "childList"
PASS mutations[21].type is "childList"
PASS mutations[1].target.tagName is "BODY"
PASS mutations[1].addedNodes.length is 1
PASS mutations[1].addedNodes[0].tagName is "P"
PASS mutations[9].target.tagName is "A"
PASS mutations[9].removedNodes[0].tagName is "FIGURE"
PASS mutations[9].removedNodes.length is 1
PASS mutations[9].addedNodes.length is 0
PASS mutations[18].target.tagName is "FIGURE"
PASS mutations[18].addedNodes.length is 1
PASS mutations[18].removedNodes.length is 0
PASS mutations[18].addedNodes[0].tagName is "DIV"
PASS successfullyParsed is true

TEST COMPLETE


Mutation records should be delivered for all parser mutations after the above script.
